+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - `font.glyph_offset.y` is now applied to underline/strikeout
+- Always use sRGB color space on macOS
 
 ### Fixed